add mpv-flex
