+++ mpv-canvas/package.json
@@ -13,6 +13,7 @@
     "@mpv-easy/easy-react": "workspace:*",
     "@mpv-easy/tool": "workspace:*",
     "@mpv-easy/react": "workspace:*",
+    "@mpv-easy/flex": "workspace:*",
     "e-color": "^0.1.3",
     "lodash-es": "^4.17.21",
     "fabric": "6.0.2",
