+++ mpv-easy-react/src/mpv-easy-es5.tsx
@@ -1,11 +1,11 @@
 import "@mpv-easy/polyfill"
 import "@mpv-easy/tool"
-// import "./mpv-easy"
+import "./mpv-easy"
 
-import React from "react"
-import { ColorBox } from './example/color-box'
-import { createRender } from "@mpv-easy/react"
+// import React from "react"
+// import { ColorBox } from "./example/color-box"
+// import { createRender } from "@mpv-easy/react"
 
-createRender({
-  showFps: true
-})(<ColorBox />)
+// createRender({
+//   showFps: true,
+// })(<ColorBox />)
