+++ mpv-easy-solid/build.ts
@@ -1,19 +1,19 @@
 import { build } from "esbuild"
 import { solidPlugin } from "esbuild-plugin-solid"
-import fs from 'fs'
+import fs from "node:fs"
 
 build({
   entryPoints: ["./src/mpv-easy-*.ts"],
   bundle: true,
   outdir: "bundle",
-  // minify: true,
+  minify: true,
 
   // TODO: solid not support node platform?
   // platform:"node",
-  charset: 'utf8',
+  charset: "utf8",
   define: {
-    "globalThis.version": '\"0.1.9\"',
-    "process.env.NODE_ENV": '\"production\"'
+    "globalThis.version": '"0.1.9"',
+    "process.env.NODE_ENV": '"production"',
   },
   plugins: [
     solidPlugin({
@@ -25,7 +25,8 @@ build({
   ],
 
   metafile: true,
-}).then(
-  r => {
-    fs.writeFileSync('./bundle/metafile.json', JSON.stringify(r.metafile));
-  }).catch(() => process.exit(1))
+})
+  .then((r) => {
+    fs.writeFileSync("./bundle/metafile.json", JSON.stringify(r.metafile))
+  })
+  .catch(() => process.exit(1))
