+++ mpv-easy-solid/src/examples/color-box.tsx
@@ -1,11 +1,9 @@
-import { getOsdSize } from "@mpv-easy/tool"
-import { Box, render, useOsdDimensions } from "@mpv-easy/solid"
-import React from "react"
+import { Box, useOsdDimensions } from "@mpv-easy/solid"
+import { COLORS } from "e-color"
+import { createSignal } from "solid-js"
 
 const row = 10
 const col = 10
-import { COLORS } from "e-color"
-import { createSignal } from "solid-js"
 
 const colorList = Object.values(COLORS).map((i) =>
   i.toString(16).padStart(6, "0"),
@@ -15,13 +13,12 @@ export function ColorBox() {
   const [count, setCount] = createSignal(0)
 
   setInterval(() => {
-    setCount(c => c + 1)
+    setCount((c) => c + 1)
   }, 100)
 
   const osd = useOsdDimensions()!
   const boxW = () => osd().w / col
   const boxH = () => osd().h / row
-  console.log(boxW, boxH, colorList.join(","))
 
   return new Array(row * col).fill(0).map((_, k) => {
     return (
