+++ mpv-easy-solid/src/examples/index.tsx
@@ -2,10 +2,9 @@ import "@mpv-easy/polyfill"
 import "@mpv-easy/tool"
 import { Box, render, effect } from "@mpv-easy/solid"
 
-
-import { SimpleBox } from './box'
-import { CounterUI } from './counter-ui'
-import { SnakeGame } from './snake'
-import { ColorBox } from './color-box'
+import { SimpleBox } from "./box"
+import { CounterUI } from "./counter-ui"
+import { SnakeGame } from "./snake"
+import { ColorBox } from "./color-box"
 
 render(() => <ColorBox />)
