+++ mpv-easy-solid/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../tsconfig.base.json",
   "compilerOptions": {
-        "jsx": "preserve",
+    "jsx": "preserve",
     "jsxImportSource": "solid-js"
   }
 }
