+++ mpv-react/package.json
@@ -19,7 +19,8 @@
     "lodash-es": "^4.17.21",
     "@mpv-easy/assdraw": "workspace:*",
     "@mpv-easy/tool": "workspace:*",
-    "@r-tui/flex": "0.1.6",
+    "@mpv-easy/flex": "workspace:*",
+    "@r-tui/flex": "file://C:/wt/r-tui/flex/dist",
     "@r-tui/share": "0.1.6",
     "e-color": "^0.1.3",
     "react-use": "^17.5.0"
