+++ mpv-react/src/ui/dropdown.tsx
@@ -2,7 +2,7 @@ import React, { useRef, useState } from "react"
 import { Button, type ButtonProps } from "./button"
 import { Box } from "./box"
 import { BaseMouseEvent, EventNameList } from "@r-tui/flex"
-import type { MpAttrs, MpDom, MpEvent, MpProps } from "../render/dom"
+import type { MpAttrs, MpDom, MpEvent, MpProps } from "@mpv-easy/flex"
 
 export type DropdownItem = {
   key: string | number
