+++ mpv-solid/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./ui"
-export * from "./render"
-export * from "./common"
-export * from './hook'
+export * from "./hook"
+export * from "./reconciler"
