--- mpv-solid/src/render/index.ts
@@ -1,3 +1,0 @@
-export * from "./reconciler"
-export * from "./flex"
-export * from "./dom"
